add method to send friend request
method creates two Friendship objects, one for the sender that has an 'accepted' value of true,
as the sender of the request implies their intention of becoming friends via the request. the other Friendship object has the default 'accepted' value, as it depends on the response
of the receiver of the request.