add Notification creation upon acceptance of friend request. refactor #accept_friend_request
Notification is created when a User accepts a friend request from the original sender. The original
sender is sent a notification letting them know of the acceptance of the friendship.

'request' variable was a collection instead of a single active record object. request is now iterated
over and each active record object in the collection has its 'accepted' value set to true. Each value
is set to true because this accounts for there being multiple of the same friend request (one user friending another user multiple times)