refactor User/Friendship assocation
Changed so that only one Friendship is created until the receiver of that Friendship accepts the
request. Upon acceptance another Friendship is created so that the relationship between both Users
is reciprocal. This also helped to remove some control flow from the #create_notification callback,
as well as allowed for the removal of a Notification creation from inside the #accept_friend_request method, so
that the callback can be utilized upon creation of the reciprocal Friendship instead.